Added home page and improved graph service
