+++ views/graph.html
@@ -7,10 +7,8 @@
     <h1>Rippled Peer Graph</h1>
     <div class="row">
       <div class="col-md-12"> The Ripple Peer Graph is a visual representation of the rippled network. </div>
-      <div class="col-md-12"> Each node is a rippled and the links between the nodes are tcp connections. </div>
+      <div class="col-md-12"> Each node is a rippled and the links between the nodes are tcp connections.  </div>
       <div class="col-md-12"> A node's size corresponds to the number of connections it has and its color indicates the software version </div>
-      <div class="col-md-12" class="disclaimer"> The Ripple Peer Dashboard is built and maintained by Ripple Labs, based on data from the Ripple Network. </div>
-      <div class="col-md-12" class="disclaimer"> Information is provided "as is" and solely as an informative resource. Data may be delayed or incorrect. </div>
     </div>
     <hr/>
     <div class='loading row' ng-show='loading'>
@@ -19,8 +17,7 @@
   </div>
 
   <div class='row' ng-hide='loading'>
-    <div class='graph'>
-    </div>
+    <div class='graph'></div>
   </div>
 
 </body>
