+++ views/home.html
@@ -1,14 +1,24 @@
-<div class='row'>
-  <h1>Rippled Peer Dashboard</h1>
-  <div class="row">
-  </div>
-  <hr/>
-  <div class='loading row' ng-show='loading'>
-    <h2>{{status}}</h2>
-  </div>
-</div>
+<!DOCTYPE html>
+<meta charset="utf-8">
 
-<div class='row' ng-hide='loading'>
+<body>
+  <div class='row-md-2'>
+    <div class='col-md-6'>
+      <h1>Rippled Peer Dashboard</h1>
+      <div class="row-md-2"> The Rippled Peer Dashboard is a summary of the rippled network </div>
+      <div class="row-md-2"> Peers are rippleds which hold tcp connections with other peers. </div>
+    </div>
+
+    <div class='loading row' ng-show='loadingGraph'>
+      <h2>{{statusGraph}}</h2>
+    </div>
+    <a href="/#/graph"> <div class='graph col-md-6'></div> </a>
+  </div>
+
+  <div class='loading row' ng-show='loadingPeers'>
+    <h2>{{statusPeers}}</h2>
+  </div>
+
   <table class='table table-striped'>
 
     <thead>
@@ -34,4 +44,5 @@
     </tr>
 
   </table>
-</div>
+
+</body>
